Added NTP service to startup
A NTP service script will run a NTP client on a custom NTP server
in order to synchronize the camera's clock. This may be useful in
cases where the camera has to use a local NTP server rather than
the one hard coded in iCamera.

To use this, simply set your desired NTP server in the `NTP_SERVER`
option in `wz_mini.conf`.