Added auto bind feature for offline / cloud-less setup (#303)
* Added auto bind feature for offline / cloud-less setup

Automatically sets bindOk=1 when WiFi settings are detected without
needing the camera to interact with the Wyze cloud. This feature
can be enabled when ENABLE_AUTO_BIND=true is set in wz_mini.conf

* Added 'WiFi Connected' audio message when QR code is valid

* Fixed the auto bind script improperly checking WiFi status

/configs isn't mounted during the init phase from the SD card and
as a result, the checks for whether bindOk=0 done by the init
script was ineffective. The script now mounts it if necessary before
checking.

* Updated autobind feature to use ENABLE_SELFHOSTED_MODE option

The feature to update the bindOk value will now use the same
`ENABLE_SELFHOSTED_MODE` option as the self-hosted iCamera patch.

This feature should only be used by advanced users in isolated
networks or when the user does not intend to use the Wyze App or
cloud services.

* Updated ENABLE_SELFHOSTED_MODE to false by default

The default should be to rely on Wyze servers. Ooops.