Update S09dnsmasq
Added logic to use the newly included dnsmasq process from Wyze on newer firmwares. First, the script will detect the newer dnsmasq process if present and kill it, next, it will determine if the newer dnsmasq binary IS present and start dnsmasq back up using our config from wz_mini. If not, it will simply launch wz_mini's build of dnsmasq. This effectively fixes the script for newer versions of the Wyze firmware.